Apply fix for crash in 1.2.6 for 1.2.7 and upcoming release
This avoids manually locating and deleting `parasiterc` and thus allows
retaining settings.
