+++ batcher/src/update/_handlers/update_1_2_7.py
@@ -1,0 +1,17 @@
+from .. import _common as update_common_
+from .. import _utils as update_utils_
+
+
+def update(data, _settings, _procedure_groups):
+  main_settings_list, _index = update_utils_.get_top_level_group_list(data, 'main')
+
+  if main_settings_list is not None:
+    actions_list, _index = update_utils_.get_child_group_list(main_settings_list, 'actions')
+
+    if actions_list is not None:
+      for index, action_dict in enumerate(actions_list):
+        action_list = action_dict['settings']
+
+        arguments_list, _index = update_utils_.get_child_group_list(action_list, 'arguments')
+
+        update_common_.update_dimension_arguments(arguments_list)
